pyproject.toml
[project]
name = "repl-nix-blueprintnavigator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.8.1",
    "numpy>=1.24.0",
    "openai>=1.54.3",
    "pandas>=1.5.3",
    "plotly>=5.13.0",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.3.0",
    "streamlit>=1.24.0",
]

[tool.uv.sources]

[tool.poetry]
name = "blueprintnavigator"
version = "0.1.0"
description = ""
authors = ["Sophia <sophia@winnie.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
